Re #6711 On unregister, move recache outside of try
Previously, the recache would never be reached if an unregister failed.
