+++ package.yaml
@@ -311,6 +311,7 @@ library:
   - Stack.Types.GHCVariant
   - Stack.Types.GhcOptionKey
   - Stack.Types.GhcOptions
+  - Stack.Types.GhcPkgExe
   - Stack.Types.GhcPkgId
   - Stack.Types.GlobalOpts
   - Stack.Types.GlobalOptsMonoid
