+++ stack.cabal
@@ -1,6 +1,6 @@
 cabal-version: 2.2
 
--- This file has been generated from package.yaml by hpack version 0.38.0.
+-- This file has been generated from package.yaml by hpack version 0.38.1.
 --
 -- see: https://github.com/sol/hpack
 
@@ -347,6 +347,7 @@ library
       Stack.Types.GHCVariant
       Stack.Types.GhcOptionKey
       Stack.Types.GhcOptions
+      Stack.Types.GhcPkgExe
       Stack.Types.GhcPkgId
       Stack.Types.GlobalOpts
       Stack.Types.GlobalOptsMonoid
